+++ torch_mlu_ext/gen_mlu_extension/mlu_custom_ext/src/custom_elu_forward_mlu.h
@@ -1,3 +1,3 @@
 #pragma once
 #include <torch/extension.h>
-torch::Tensor elu_forward_mlu(torch::Tensor input, float alpha);
+torch::Tensor elu_forward_mlu(torch::Tensor input, double alpha);
