+++ torch_mlu_ext/gen_mlu_extension/mlu_custom_ext/src/custom_matrix_scalar_mul_mlu.h
@@ -1,3 +1,3 @@
 #pragma once
 #include <torch/extension.h>
-torch::Tensor matrix_scalar_mul_mlu(torch::Tensor A, float s);
+torch::Tensor matrix_scalar_mul_mlu(torch::Tensor A, double s);
