+++ torch_mlu_ext/gen_mlu_extension/mlu_custom_ext/src/custom_mse_mlu.cpp
@@ -21,8 +21,8 @@ torch::Tensor mse_mlu(torch::Tensor predictions, torch::Tensor targets) {
     auto size = predictions_contiguous.numel();
     auto out = at::zeros_like(predictions_contiguous);
     
-    const int block_size = 256;
-    const int num_blocks = (size + block_size - 1) / block_size;
+    const int64_t block_size = 256;
+    const int64_t num_blocks = (size + block_size - 1) / block_size;
     
     auto out_contiguous = torch_mlu::cnnl_contiguous(out);
     auto out_impl = getMluTensorImpl(out_contiguous);
