+++ torch_mlu_ext/gen_mlu_extension/mlu_custom_ext/src/custom_product_reduction_mlu.h
@@ -1,3 +1,3 @@
 #pragma once
 #include <torch/extension.h>
-torch::Tensor product_reduction_mlu(torch::Tensor input, int reduction_dim);
+torch::Tensor product_reduction_mlu(torch::Tensor input, int64_t reduction_dim);
