+++ torch_mlu_ext/gen_mlu_extension/mlu_custom_ext/src/custom_sum_reduction_mlu.h
@@ -1,3 +1,3 @@
 #pragma once
 #include <torch/extension.h>
-torch::Tensor sum_reduction_mlu(torch::Tensor x, int dim);
+torch::Tensor sum_reduction_mlu(torch::Tensor x, int64_t dim);
