+++ torch_mlu_ext/gen_mlu_extension/mlu_custom_ext/src/custom_triplet_margin_loss_mlu.h
@@ -1,3 +1,3 @@
 #pragma once
 #include <torch/extension.h>
-torch::Tensor triplet_margin_loss_mlu(torch::Tensor anchor, torch::Tensor positive, torch::Tensor negative, float margin);
+torch::Tensor triplet_margin_loss_mlu(torch::Tensor anchor, torch::Tensor positive, torch::Tensor negative, double margin);
